// src/data/node-types/programme-episodes/types/ProgrammeEpisodeHasMainImageRelationship.ts
import type {DataRelationshipType} from "../../../types/DataRelationshipType"
import type {ProgrammeEpisodeNode} from "./ProgrammeEpisodeNode"
import type {ImageNode} from "../../images/types/ImageNode"

export type ProgrammeEpisodeHasMainImageRelationship = {
    id: number
    name: DataRelationshipType.PROGRAMME_EPISODE_HAS_MAIN_IMAGE
    source_node: ProgrammeEpisodeNode
    partner_node: ImageNode
    created_at: string
    updated_at: string
}